+++ packages/utils/src/enums/token.enums.ts
@@ -1,4 +1,5 @@
 export enum FromStringToTokenError {
   FractionalMoreThan8Decimals,
   InvalidFormat,
+  FractionalTooManyDecimals,
 }